+++ libs/core/platform.h
@@ -1,1 +1,3 @@
-// leave empty
+#define PXT_GC_THREAD_LIST 1
+
+#define PXT_IN_ISR() false
